refactor(AppBar): move to /utils
